refactor(repo): split nx-deploy analysis out of github lib
