--- libs/shared/util/github/package.json
@@ -1,14 +1,0 @@
-{
-  "name": "@codeware/shared-util-github",
-  "version": "0.0.1",
-  "private": true,
-  "dependencies": {
-    "@actions/core": "^1.11.1",
-    "@actions/exec": "^1.1.1",
-    "@actions/github": "^6.0.0",
-    "@nx/devkit": "^23.1.0",
-    "@octokit/request-error": "^5.0.1",
-    "http-status-codes": "^2.3.0",
-    "zod": "^3.25.76"
-  }
-}
