--- libs/shared/util/github/src/internal.ts
@@ -1,20 +1,0 @@
-// For internal use only!
-
-export { analyzeAppsToDeploy } from './lib/_actions-internal/analyze-apps-to-deploy';
-export {
-  type DeployableApp,
-  DeployableAppSchema
-} from './lib/_actions-internal/deployable-app.schema';
-export {
-  AppDeploymentDetailsSchema,
-  AppDetailsSchema,
-  type AppDeploymentDetails,
-  type AppDetails
-} from './lib/_actions-internal/app-deployment-details.schema';
-export { getAppName } from './lib/_actions-internal/get-app-name';
-export { getNxApps } from './lib/_actions-internal/get-nx-apps';
-export { getNxProject } from './lib/_actions-internal/get-nx-project';
-export {
-  type Environment,
-  EnvironmentSchema
-} from './lib/_actions-internal/environment.schema';
