+++ libs/shared/util/nx-deploy/project.json
@@ -1,0 +1,7 @@
+{
+  "name": "shared-util-nx-deploy",
+  "$schema": "../../../../node_modules/nx/schemas/project-schema.json",
+  "sourceRoot": "libs/shared/util/nx-deploy/src",
+  "projectType": "library",
+  "tags": ["scope:shared", "type:util"]
+}
