+++ libs/shared/util/nx-deploy/src/index.ts
@@ -1,0 +1,15 @@
+export { analyzeAppsToDeploy } from './lib/analyze-apps-to-deploy';
+export {
+  type DeployableApp,
+  DeployableAppSchema
+} from './lib/deployable-app.schema';
+export {
+  AppDeploymentDetailsSchema,
+  AppDetailsSchema,
+  type AppDeploymentDetails,
+  type AppDetails
+} from './lib/app-deployment-details.schema';
+export { getAppName } from './lib/get-app-name';
+export { getNxApps } from './lib/get-nx-apps';
+export { getNxProject } from './lib/get-nx-project';
+export { type Environment, EnvironmentSchema } from './lib/environment.schema';
