+++ libs/shared/util/nx-deploy/src/lib/app-deployment-details.schema.ts
