+++ libs/shared/util/nx-deploy/src/lib/get-app-name.ts
