+++ libs/shared/util/nx-deploy/src/lib/get-nx-apps.ts
