+++ libs/shared/util/nx-deploy/src/lib/get-nx-project.ts
