+++ libs/shared/util/nx-deploy/vite.config.mts
@@ -1,0 +1,21 @@
+import { nxViteTsPaths } from '@nx/vite/plugins/nx-tsconfig-paths.plugin';
+import { defineConfig } from 'vite';
+
+export default defineConfig({
+  root: __dirname,
+  cacheDir: '../../../../node_modules/.vite/libs/shared/util/nx-deploy',
+  plugins: [nxViteTsPaths()],
+  test: {
+    name: 'shared-util-nx-deploy',
+    watch: false,
+    globals: true,
+    environment: 'node',
+    include: ['src/**/*.{test,spec}.{js,mjs,cjs,ts,mts,cts,jsx,tsx}'],
+    reporters: ['default'],
+    coverage: {
+      reportsDirectory: '../../../../coverage/libs/shared/util/nx-deploy',
+      provider: 'v8'
+    },
+    passWithNoTests: true
+  }
+});
