+++ packages/fly-build-action/src/lib/schemas/action-inputs.schema.ts
@@ -2,7 +2,7 @@ import {
   type AppDeploymentDetails,
   AppDetailsSchema,
   DeployableAppSchema
-} from '@codeware/shared/util/github/internal';
+} from '@codeware/shared/util/nx-deploy';
 import { z } from 'zod';
 
 export const ActionInputsSchema = z.object({
