+++ packages/fly-build-action/src/lib/utils/run-build-apps.ts
@@ -1,9 +1,6 @@
 import * as core from '@actions/core';
 import { Fly } from '@codeware/fly-node';
-import {
-  type Environment,
-  getAppName
-} from '@codeware/shared/util/github/internal';
+import { type Environment, getAppName } from '@codeware/shared/util/nx-deploy';
 
 import type { BuildConfig } from '../schemas/build-config.schema';
