+++ packages/fly-deployment-action/src/lib/schemas/action-outputs.schema.ts
@@ -1,4 +1,4 @@
-import { EnvironmentSchema } from '@codeware/shared/util/github/internal';
+import { EnvironmentSchema } from '@codeware/shared/util/nx-deploy';
 import { z } from 'zod';
 
 export const ActionSchema = z.enum(['deploy', 'destroy', 'failed']);
