+++ packages/fly-deployment-action/src/lib/schemas/context.schema.ts
@@ -1,7 +1,7 @@
 import {
   type Environment,
   EnvironmentSchema
-} from '@codeware/shared/util/github/internal';
+} from '@codeware/shared/util/nx-deploy';
 import { z } from 'zod';
 
 type PrettyPartial<T> = {
